renamed i_control to control_interface
