+++ pythoncode/src/osm_configurator/control/control_interface.py
